add USBDriver_Impl::DeviceExists so we can see if a device exists before trying to instantiate it

+++ src/io/USBDriver.h
@@ -17,7 +17,7 @@ public:
 	USBDriver();
 	virtual ~USBDriver();
 
-	bool OpenInternal( int iVendorID, int iProductID );
+	bool OpenInternal( short iVendorID, short iProductID );
 
 	virtual bool Open();
 	virtual void Close();
